Change int8 to int16 so describe() and other functions work

+++ api/python/vds1022/vds1022.py
@@ -447,7 +447,7 @@ class Frame:
     def _points(self):
         buf = self.buffer
         if isinstance(buf, array) and buf.itemsize == 1:
-            buf = np.frombuffer(buf, np.int8)
+            buf = np.frombuffer(buf, np.int16)
             buf.clip(ADC_MIN, ADC_MAX, out=buf)
             self.buffer = buf
         return buf
